Sistemati Unit Test

+++ ArticoliWebService/Models/Iva.cs
@@ -1,5 +1,6 @@
 using System.ComponentModel.DataAnnotations;
 using System.Collections.Generic;
+using System;
 
 namespace ArticoliWebService.Models
 {
@@ -11,7 +12,7 @@ namespace ArticoliWebService.Models
         [MaxLength(80, ErrorMessage = "Il numero massimo di caratteri è 80")]
         public string Descrizione { get; set; }
         [Required]
-        public int Aliquota { get; set; }
+        public Int16 Aliquota { get; set; }
         public virtual ICollection<Articoli> Articoli { get; set; }
     }
 }
